README and Default Log Location Update
Updated the README to include Windows 11 and further GPO guidance such as suggesting the use of Rob Willis' blogpost on setting up the GPO.

Updated the PowerShell script to use a log filename that corresponds to the new script name and not the script name used by Rob Willis. Also updated the description in the script to reference the new log filename and changes made to identify that this works on Windows 11. i.e. 'Windows 10/11 or Server 12/16/19/22' vs 'Windows 10/Server 12/16/19/22'